Drasil.Database.SearchTools

Documentation

data TermAbbr

Constructors

TermAbbr

Fields

longForm :: NP

shortForm :: Maybe String

termResolve :: (NP -> Maybe String -> c) -> ChunkDB -> UID -> c

Search for a chunk that is an instance of Idea and return its "term" and abbreviation, erroring out if it doesn't exist.

termResolve' :: ChunkDB -> UID -> TermAbbr

Find a chunk's "term" and abbreviation, erroring out if it doesn't exist.

data DomDefn

Constructors

DomDefn

Fields

domain :: [UID]

definition :: Sentence

defResolve :: ([UID] -> Sentence -> c) -> ChunkDB -> UID -> c

Looks up a UID in all tables with concepts from the ChunkDB. If nothing is found, an error is thrown.

defResolve' :: ChunkDB -> UID -> DomDefn

findAllConcInsts :: ChunkDB -> [ConceptInstance]

findAllCitations :: ChunkDB -> [Citation]

findAllLabelledContent :: ChunkDB -> [LabelledContent]