Theory.Drasil

Documentation

class HasInputs c where

Members of this class may have inputs.

Methods

inputs :: Lens' c [(QuantityDict, Maybe (RealInterval Expr Expr))]

Provides a Lens that holds a QuantityDict and maybe constraints.

class HasOutput c where

Members of this class may have outputs.

Methods

output :: Lens' c QuantityDict

Provides a Lens that holds a QuantityDict for output.

out_constraints :: Lens' c [RealInterval Expr Expr]

Provides a Lens that holds constraints on the output.

data DataDefinition

A data definition is a QDefinition that may have additional notes: the scope, any references, maybe a derivation, a label (ShortName), a reference address, and other notes (Sentences).

dd :: QDefinition -> [Reference] -> Maybe Derivation -> String -> [Sentence] -> DataDefinition

Smart constructor for data definitions.

ddNoRefs :: QDefinition -> Maybe Derivation -> String -> [Sentence] -> DataDefinition

Smart constructor for data definitions with no references.

qdFromDD :: DataDefinition -> QDefinition

Extracts the QDefinition from a DataDefinition.

data GenDefn

A general definition is a ModelKind that may have units, a derivation, references, a shortname, a reference address, and notes.

gd :: IsUnit u => ModelKinds -> Maybe u -> Maybe Derivation -> [Reference] -> String -> [Sentence] -> GenDefn

Smart constructor for general definitions derived from ModelKinds.

gdNoRefs :: IsUnit u => ModelKinds -> Maybe u -> Maybe Derivation -> String -> [Sentence] -> GenDefn

Smart constructor for general definitions with no references, derived from ModelKinds.

gd' :: IsUnit u => UID -> ModelKinds -> Maybe u -> Maybe Derivation -> [Reference] -> String -> [Sentence] -> GenDefn

Smart constructor for general definitions.

gdNoRefs' :: IsUnit u => UID -> ModelKinds -> Maybe u -> Maybe Derivation -> String -> [Sentence] -> GenDefn

Smart constructor for general definitions with no references.

getEqModQdsFromGd :: [GenDefn] -> [QDefinition]

Grab all related QDefinitions from a list of general definitions.

data ModelKinds

Models can be of different kinds:

Equational models contain QDefinitions,

Differential equation models contain RelationConcepts,

Other models contain RelationConcepts.

Constructors

EquationalModel QDefinition

DEModel RelationConcept

OthModel RelationConcept

getEqModQds :: [ModelKinds] -> [QDefinition]

Extract a list of QDefinitions from a list of ModelKinds.

data InstanceModel

An instance model is a ModelKind that may have specific inputs, outputs, and output constraints. It also has attributes like references, derivation, labels (ShortName), reference address, and notes.

im :: ModelKinds -> Inputs -> Output -> OutputConstraints -> [Reference] -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with everything defined.

imNoDeriv :: ModelKinds -> Inputs -> Output -> OutputConstraints -> [Reference] -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with no derivation.

imNoRefs :: ModelKinds -> Inputs -> Output -> OutputConstraints -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with no references.

imNoDerivNoRefs :: ModelKinds -> Inputs -> Output -> OutputConstraints -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with no derivations or references.

qwUC :: (Quantity q, MayHaveUnit q) => q -> Input

For building a quantity with no constraint.

qwC :: (Quantity q, MayHaveUnit q) => q -> RealInterval Expr Expr -> Input

For building a quantity with a constraint.

getEqModQdsFromIm :: [InstanceModel] -> [QDefinition]

Grab all related QDefinitions from a list of instance models.

class Theory t where

Theories are the basis for building models with context, spaces, quantities, operations, invariants, etc.

Methods

valid_context :: Lens' t [TheoryModel]

spaces :: Lens' t [SpaceDefn]

quantities :: Lens' t [QuantityDict]

operations :: Lens' t [ConceptChunk]

defined_quant :: Lens' t [QDefinition]

invariants :: Lens' t [Relation]

defined_fun :: Lens' t [QDefinition]

data TheoryModel

A TheoryModel is a collection of:

tUid - a UID,

con - a ConceptChunk,

vctx - definition context (TheoryModels),

spc - type definitions (SpaceDefns),

quan - quantities (QuantityDicts),

ops - operations (ConceptChunks),

defq - definitions (QDefinitions),

invs - invariants (Relations),

dfun - defined functions (QDefinitions),

ref - accompanying references (References),

lb - a label (SpaceDefn),

ra - reference address (SpaceDefn),

notes - additional notes (Sentences).

Right now, neither the definition context (vctx) nor the spaces (spc) are ever defined.

tm :: (Quantity q, MayHaveUnit q, Concept c) => ModelKinds -> [q] -> [c] -> [QDefinition] -> [Relation] -> [QDefinition] -> [Reference] -> String -> [Sentence] -> TheoryModel

Constructor for theory models.

tmNoRefs :: (Quantity q, MayHaveUnit q, Concept c) => ModelKinds -> [q] -> [c] -> [QDefinition] -> [Relation] -> [QDefinition] -> String -> [Sentence] -> TheoryModel

Constructor for theory models with no references.

tm' :: (Quantity q, MayHaveUnit q, Concept c) => UID -> ModelKinds -> [q] -> [c] -> [QDefinition] -> [Relation] -> [QDefinition] -> [Reference] -> String -> [Sentence] -> TheoryModel

Constructor for theory models. Must have a source. Uses the shortname of the reference address.

tmNoRefs' :: (Quantity q, MayHaveUnit q, Concept c) => UID -> ModelKinds -> [q] -> [c] -> [QDefinition] -> [Relation] -> [QDefinition] -> String -> [Sentence] -> TheoryModel

Constructor for theory models. Uses the shortname of the reference address.