Theory.Drasil.Components.Derivation

Description

For deriving equations in examples.

Type

data Derivation

Derivations are an ordered list of sentences and expressions. They are rendered in order as paragraphs and equation blocks to display the derivation.

Constructors

Derivation Sentence [Sentence]

Class

class MayHaveDerivation c where

A class that might have a Derivation.

Methods

derivations :: Lens' c (Maybe Derivation)

Provides a Lens to a possible derivation.

Functions

mkDeriv :: Sentence -> [Sentence] -> Derivation

Smart constructor for creating a Derivation.

mkDerivName :: Sentence -> [Sentence] -> Derivation

Similar to mkDeriv, but prepends "Detailed derivation of" to the header.

mkDerivNoHeader :: [Sentence] -> Derivation

Similar to mkDeriv, but without a header Sentence.