Theory.Drasil.InstanceModel

Documentation

data InstanceModel

An instance model is a ModelKind that may have specific inputs, outputs, and output constraints. It also has attributes like references, derivation, labels (ShortName), reference address, and notes.

im :: ModelKind -> Inputs -> Output -> OutputConstraints -> [Reference] -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with everything defined.

imNoDeriv :: ModelKind -> Inputs -> Output -> OutputConstraints -> [Reference] -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no derivation.

imNoRefs :: ModelKind -> Inputs -> Output -> OutputConstraints -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no references.

imNoDerivNoRefs :: ModelKind -> Inputs -> Output -> OutputConstraints -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no derivations or references.

getEqModQdsFromIm :: [InstanceModel] -> [QDefinition]

Grab all related QDefinitions from a list of instance models.

qwUC :: (Quantity q, MayHaveUnit q) => q -> Input

For building a quantity with no constraint.

qwC :: (Quantity q, MayHaveUnit q) => q -> RealInterval Expr Expr -> Input

For building a quantity with a constraint.