Theory.Drasil.ModelKinds

Documentation

data ModelKind

ModelKinds carrier, used to carry commonly overwritten information from the IMsTMsGDs.

Constructors

MK

Fields

_mk :: ModelKinds

_mkUid :: UID

_mkTerm :: NP

data ModelKinds

Models can be of different kinds:

DEModels represent differential equations as RelationConcepts

EquationalConstraints represent invariants that will hold in a system of equations.

EquationalModels represent quantities that are calculated via a single definition/QDefinition.

EquationalRealms represent MultiDefns; quantities that may be calculated using any one of many DefiningExprs (e.g., 'x = A = ... = Z')

OthModels are placeholders for models. No new OthModels should be created, they should be using one of the other kinds.

Constructors

DEModel RelationConcept

EquationalConstraints ConstraintSet

EquationalModel QDefinition

EquationalRealm MultiDefn

OthModel RelationConcept

deModel :: UID -> NP -> RelationConcept -> ModelKind

Smart constructor for DEModels

equationalConstraints :: UID -> NP -> ConstraintSet -> ModelKind

Smart constructor for EquationalConstraints

equationalModel :: UID -> NP -> QDefinition -> ModelKind

Smart constructor for EquationalModels

equationalRealm :: UID -> NP -> MultiDefn -> ModelKind

Smart constructor for EquationalRealms

othModel :: UID -> NP -> RelationConcept -> ModelKind

Smart constructor for OthModels

deModel' :: RelationConcept -> ModelKind

Smart constructor for DEModels, deriving UID+Term from the RelationConcept

equationalConstraints' :: ConstraintSet -> ModelKind

Smart constructor for EquationalConstraints, deriving UID+Term from the ConstraintSet

equationalModel' :: QDefinition -> ModelKind

Smart constructor for EquationalModels, deriving UID+Term from the QDefinition

equationalRealm' :: MultiDefn -> ModelKind

Smart constructor for EquationalRealms

othModel' :: RelationConcept -> ModelKind

Smart constructor for OthModels, deriving UID+Term from the RelationConcept

equationalModelU :: UID -> QDefinition -> ModelKind

Smart constructor for EquationalModels, deriving Term from the QDefinition

equationalModelN :: NP -> QDefinition -> ModelKind

Smart constructor for EquationalModels, deriving UID from the QDefinition

equationalRealmU :: UID -> MultiDefn -> ModelKind

Smart constructor for EquationalRealms

equationalRealmN :: NP -> MultiDefn -> ModelKind

Smart constructor for EquationalRealms, deriving UID from the MultiDefn

setMk :: ModelKinds -> Setter' RelationConcept a -> Setter' ConstraintSet a -> Setter' QDefinition a -> Setter' MultiDefn a -> a -> ModelKinds

Map into internal representations of ModelKinds

elimMk :: Getter RelationConcept a -> Getter ConstraintSet a -> Getter QDefinition a -> Getter MultiDefn a -> ModelKinds -> a

Retrieve internal data from ModelKinds

lensMk :: forall a. Lens' RelationConcept a -> Lens' ConstraintSet a -> Lens' QDefinition a -> Lens' MultiDefn a -> Lens' ModelKinds a

Make a Lens for ModelKinds.

getEqModQds :: [ModelKind] -> [QDefinition]

Extract a list of QDefinitions from a list of ModelKinds.