Theory.Drasil.MultiDefn

Documentation

data MultiDefn

MultiDefns are QDefinition factories, used for showing one or more ways we can define a QDefinition.

data DefiningExpr

DefiningExpr are the data that make up a (quantity) definition, namely the description, the defining (rhs) expression and the context domain(s). These are meant to be alternate but equivalent definitions for a single concept.

mkMultiDefn :: UID -> QuantityDict -> Sentence -> NonEmpty DefiningExpr -> MultiDefn

Smart constructor for MultiDefns, does nothing special at the moment

mkMultiDefnForQuant :: QuantityDict -> Sentence -> NonEmpty DefiningExpr -> MultiDefn

Smart constructor for MultiDefns defining UIDs using that of the QuantityDict

mkDefiningExpr :: UID -> [UID] -> Sentence -> Expr -> DefiningExpr

Smart constructor for DefiningExprs

multiDefnGenQD :: MultiDefn -> DefiningExpr -> QDefinition

Convert MultiDefns into QDefinitions via a specific DefiningExpr

multiDefnGenQDByUID :: MultiDefn -> UID -> QDefinition

Convert MultiDefns into QDefinitions via a specific DefiningExpr (by UID)