Theory.Drasil

Documentation

class HasInputs c where

Members of this class may have inputs.

Methods

inputs :: Lens' c [(QuantityDict, Maybe (RealInterval Expr Expr))]

Provides a Lens that holds a QuantityDict and maybe constraints.

class HasOutput c where

Members of this class may have outputs.

Methods

output :: Lens' c QuantityDict

Provides a Lens that holds a QuantityDict for output.

out_constraints :: Lens' c [RealInterval Expr Expr]

Provides a Lens that holds constraints on the output.

data ConstraintSet

ConstraintSets are sets of invariants that always hold for underlying domains.

mkConstraintSet :: ConceptChunk -> NonEmpty Relation -> ConstraintSet

Smart constructor for building ConstraintSets

data DataDefinition

A data definition is a QDefinition that may have additional notes: the scope, any references (as DecRefs), maybe a derivation, a label (ShortName), a reference address, and other notes (Sentences).

dd :: QDefinition -> [DecRef] -> Maybe Derivation -> String -> [Sentence] -> DataDefinition

Smart constructor for data definitions.

ddNoRefs :: QDefinition -> Maybe Derivation -> String -> [Sentence] -> DataDefinition

Smart constructor for data definitions with no references.

qdFromDD :: DataDefinition -> QDefinition

Extracts the QDefinition from a DataDefinition.

data GenDefn

A general definition is a ModelKind that may have units, a derivation, references (as DecRefs), a shortname, a reference address, and notes.

gd :: IsUnit u => ModelKind -> Maybe u -> Maybe Derivation -> [DecRef] -> String -> [Sentence] -> GenDefn

Smart constructor for general definitions.

gdNoRefs :: IsUnit u => ModelKind -> Maybe u -> Maybe Derivation -> String -> [Sentence] -> GenDefn

Smart constructor for general definitions with no references.

getEqModQdsFromGd :: [GenDefn] -> [QDefinition]

Grab all related QDefinitions from a list of general definitions.

data MultiDefn

MultiDefns are QDefinition factories, used for showing one or more ways we can define a QDefinition.

data DefiningExpr

DefiningExpr are the data that make up a (quantity) definition, namely the description, the defining (rhs) expression and the context domain(s). These are meant to be alternate but equivalent definitions for a single concept.

mkMultiDefn :: UID -> QuantityDict -> Sentence -> NonEmpty DefiningExpr -> MultiDefn

Smart constructor for MultiDefns, does nothing special at the moment

mkMultiDefnForQuant :: QuantityDict -> Sentence -> NonEmpty DefiningExpr -> MultiDefn

Smart constructor for MultiDefns defining UIDs using that of the QuantityDict

mkDefiningExpr :: UID -> [UID] -> Sentence -> Expr -> DefiningExpr

Smart constructor for DefiningExprs

multiDefnGenQD :: MultiDefn -> DefiningExpr -> QDefinition

Convert MultiDefns into QDefinitions via a specific DefiningExpr

multiDefnGenQDByUID :: MultiDefn -> UID -> QDefinition

Convert MultiDefns into QDefinitions via a specific DefiningExpr (by UID)

data ModelKind

ModelKinds carrier, used to carry commonly overwritten information from the IMsTMsGDs.

deModel :: UID -> NP -> RelationConcept -> ModelKind

Smart constructor for DEModels

equationalConstraints :: UID -> NP -> ConstraintSet -> ModelKind

Smart constructor for EquationalConstraints

equationalModel :: UID -> NP -> QDefinition -> ModelKind

Smart constructor for EquationalModels

equationalRealm :: UID -> NP -> MultiDefn -> ModelKind

Smart constructor for EquationalRealms

othModel :: UID -> NP -> RelationConcept -> ModelKind

Smart constructor for OthModels

deModel' :: RelationConcept -> ModelKind

Smart constructor for DEModels, deriving UID+Term from the RelationConcept

equationalConstraints' :: ConstraintSet -> ModelKind

Smart constructor for EquationalConstraints, deriving UID+Term from the ConstraintSet

equationalModel' :: QDefinition -> ModelKind

Smart constructor for EquationalModels, deriving UID+Term from the QDefinition

equationalRealm' :: MultiDefn -> ModelKind

Smart constructor for EquationalRealms, deriving UID+Term from the MultiDefn

othModel' :: RelationConcept -> ModelKind

Smart constructor for OthModels, deriving UID+Term from the RelationConcept

equationalModelU :: UID -> QDefinition -> ModelKind

Smart constructor for EquationalModels, deriving Term from the QDefinition

equationalModelN :: NP -> QDefinition -> ModelKind

Smart constructor for EquationalModels, deriving UID from the QDefinition

equationalRealmU :: UID -> MultiDefn -> ModelKind

Smart constructor for EquationalRealms

equationalRealmN :: NP -> MultiDefn -> ModelKind

Smart constructor for EquationalRealms, deriving UID from the MultiDefn

data InstanceModel

An instance model is a ModelKind that may have specific inputs, outputs, and output constraints. It also has attributes like references, derivation, labels (ShortName), reference address, and notes.

im :: ModelKind -> Inputs -> Output -> OutputConstraints -> [DecRef] -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with everything defined.

imNoDeriv :: ModelKind -> Inputs -> Output -> OutputConstraints -> [DecRef] -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no derivation.

imNoRefs :: ModelKind -> Inputs -> Output -> OutputConstraints -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no references.

imNoDerivNoRefs :: ModelKind -> Inputs -> Output -> OutputConstraints -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no derivations or references.

qwUC :: (Quantity q, MayHaveUnit q) => q -> Input

For building a quantity with no constraint.

qwC :: (Quantity q, MayHaveUnit q) => q -> RealInterval Expr Expr -> Input

For building a quantity with a constraint.

getEqModQdsFromIm :: [InstanceModel] -> [QDefinition]

Grab all related QDefinitions from a list of instance models.

class Theory t where

Theories are the basis for building models with context, spaces, quantities, operations, invariants, etc.

Methods

valid_context :: Lens' t [TheoryModel]

spaces :: Lens' t [SpaceDefn]

quantities :: Lens' t [QuantityDict]

operations :: Lens' t [ConceptChunk]

defined_quant :: Lens' t [QDefinition]

invariants :: Lens' t [ModelExpr]

defined_fun :: Lens' t [QDefinition]

data TheoryModel

A TheoryModel is a collection of:

tUid - a UID,

con - a ConceptChunk,

vctx - definition context (TheoryModels),

spc - type definitions (SpaceDefns),

quan - quantities (QuantityDicts),

ops - operations (ConceptChunks),

defq - definitions (QDefinitions),

invs - invariants (DisplayExprs),

dfun - defined functions (QDefinitions),

ref - accompanying references (DecRefs),

lb - a label (SpaceDefn),

ra - reference address (SpaceDefn),

notes - additional notes (Sentences).

Right now, neither the definition context (vctx) nor the spaces (spc) are ever defined.

tm :: (Quantity q, MayHaveUnit q, Concept c) => ModelKind -> [q] -> [c] -> [QDefinition] -> [ModelExpr] -> [QDefinition] -> [DecRef] -> String -> [Sentence] -> TheoryModel

Constructor for theory models. Must have a source. Uses the shortname of the reference address.

tmNoRefs :: (Quantity q, MayHaveUnit q, Concept c) => ModelKind -> [q] -> [c] -> [QDefinition] -> [ModelExpr] -> [QDefinition] -> String -> [Sentence] -> TheoryModel

Constructor for theory models. Uses the shortname of the reference address.