Theory.Drasil

Description

Re-export many things to simplify external use.

Classes

class HasInputs c where

Members of this class may have inputs.

Methods

inputs :: Lens' c [(DefinedQuantityDict, Maybe (RealInterval Expr Expr))]

Provides a Lens that holds a DefinedQuantityDict and maybe constraints.

class HasOutput c where

Members of this class may have outputs.

Methods

output :: Getter c DefinedQuantityDict

Provides a Getter that holds a DefinedQuantityDict for output.

out_constraints :: Getter c [RealInterval Expr Expr]

Provides a Getter that holds constraints on the output.

Constraint Sets

data ConstraintSet e

ConstraintSets are sets of invariants that always hold for underlying domains.

mkConstraintSet :: ConceptChunk -> NonEmpty e -> ConstraintSet e

Smart constructor for building ConstraintSets

Data Definitions

data DataDefinition

A data definition is a QDefinition that may have additional notes: the scope, any references (as DecRefs), maybe a derivation, a label (ShortName), a reference address, and other notes (Sentences).

ddE :: SimpleQDef -> [DecRef] -> Maybe Derivation -> String -> [Sentence] -> DataDefinition

Smart constructor for data definitions.

ddENoRefs :: SimpleQDef -> Maybe Derivation -> String -> [Sentence] -> DataDefinition

Smart constructor for data definitions with no references.

ddME :: ModelQDef -> [DecRef] -> Maybe Derivation -> String -> [Sentence] -> DataDefinition

Smart constructor for data definitions.

ddMENoRefs :: ModelQDef -> Maybe Derivation -> String -> [Sentence] -> DataDefinition

Smart constructor for data definitions with no references.

qdFromDD :: DataDefinition -> Either SimpleQDef ModelQDef

Extracts the 'QDefinition e' from a DataDefinition.

qdEFromDD :: DataDefinition -> Maybe SimpleQDef

General Definitions

data GenDefn

A general definition is a ModelKind that may have units, a derivation, references (as DecRefs), a shortname, a reference address, and notes.

gd :: IsUnit u => ModelKind ModelExpr -> Maybe u -> Maybe Derivation -> [DecRef] -> String -> [Sentence] -> GenDefn

Smart constructor for general definitions.

gdNoRefs :: IsUnit u => ModelKind ModelExpr -> Maybe u -> Maybe Derivation -> String -> [Sentence] -> GenDefn

Smart constructor for general definitions with no references.

getEqModQdsFromGd :: [GenDefn] -> [ModelQDef]

Grab all related QDefinitions from a list of general definitions.

MultiDefn

data MultiDefn e

MultiDefns are QDefinition factories, used for showing one or more ways we can define a QDefinition.

data DefiningExpr e

DefiningExpr are the data that make up a (quantity) definition, namely the description, the defining (rhs) expression and the context domain(s). These are meant to be alternate but equivalent definitions for a single concept.

mkMultiDefn :: String -> DefinedQuantityDict -> Sentence -> NonEmpty (DefiningExpr e) -> MultiDefn e

Smart constructor for MultiDefns, does nothing special at the moment. First argument is the Space to become a UID.

mkMultiDefnForQuant :: DefinedQuantityDict -> Sentence -> NonEmpty (DefiningExpr e) -> MultiDefn e

Smart constructor for MultiDefns defining UIDs using that of the DefinedQuantityDict.

mkDefiningExpr :: String -> [UID] -> Sentence -> e -> DefiningExpr e

Smart constructor for DefiningExprs.

multiDefnGenQD :: MultiDefn e -> DefiningExpr e -> QDefinition e

Convert MultiDefns into QDefinitions via a specific DefiningExpr.

multiDefnGenQDByUID :: MultiDefn e -> UID -> QDefinition e

Convert MultiDefns into QDefinitions via a specific DefiningExpr (by UID).

ModelKinds

data ModelKind e

ModelKinds carrier, used to carry commonly overwritten information from the IMsTMsGDs.

newDEModel :: String -> NP -> DifferentialModel -> ModelKind e

Smart constructor for NewDEModels

deModel :: String -> NP -> RelationConcept -> ModelKind e

Smart constructor for DEModels

equationalConstraints :: String -> NP -> ConstraintSet e -> ModelKind e

Smart constructor for EquationalConstraints

equationalModel :: String -> NP -> QDefinition e -> ModelKind e

Smart constructor for EquationalModels

equationalRealm :: String -> NP -> MultiDefn e -> ModelKind e

Smart constructor for EquationalRealms

othModel :: String -> NP -> RelationConcept -> ModelKind Expr

Smart constructor for OthModels

newDEModel' :: DifferentialModel -> ModelKind e

Smart constructor for NewDEModels, deriving UID+Term from the DifferentialModel

deModel' :: RelationConcept -> ModelKind e

Smart constructor for DEModels, deriving UID+Term from the RelationConcept

equationalConstraints' :: ConstraintSet e -> ModelKind e

Smart constructor for EquationalConstraints, deriving UID+Term from the ConstraintSet

equationalModel' :: QDefinition e -> ModelKind e

Smart constructor for EquationalModels, deriving UID+Term from the QDefinition

equationalRealm' :: MultiDefn e -> ModelKind e

Smart constructor for EquationalRealms, deriving UID+Term from the MultiDefn

othModel' :: RelationConcept -> ModelKind e

Smart constructor for OthModels, deriving UID+Term from the RelationConcept

equationalModelU :: String -> QDefinition e -> ModelKind e

Smart constructor for EquationalModels, deriving Term from the QDefinition

equationalModelN :: NP -> QDefinition e -> ModelKind e

Smart constructor for EquationalModels, deriving UID from the QDefinition

equationalRealmU :: String -> MultiDefn e -> ModelKind e

Smart constructor for EquationalRealms

equationalRealmN :: NP -> MultiDefn e -> ModelKind e

Smart constructor for EquationalRealms, deriving UID from the MultiDefn

Instance Models

data InstanceModel

An instance model is a ModelKind that may have specific inputs, outputs, and output constraints. It also has attributes like references, derivation, labels (ShortName), reference address, and notes.

im :: ModelKind Expr -> Inputs -> Output -> OutputConstraints -> [DecRef] -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with everything defined.

imNoDeriv :: ModelKind Expr -> Inputs -> Output -> OutputConstraints -> [DecRef] -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no derivation.

imNoRefs :: ModelKind Expr -> Inputs -> Output -> OutputConstraints -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no references.

imNoDerivNoRefs :: ModelKind Expr -> Inputs -> Output -> OutputConstraints -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no derivations or references.

qwUC :: (Quantity q, MayHaveUnit q, Concept q) => q -> Input

For building a quantity with no constraint.

qwC :: (Quantity q, MayHaveUnit q, Concept q) => q -> RealInterval Expr Expr -> Input

For building a quantity with a constraint.

getEqModQdsFromIm :: [InstanceModel] -> [SimpleQDef]

Grab all related QDefinitions from a list of instance models.

Theory Models

class Theory t where

Theories are the basis for building models with context, spaces, quantities, operations, invariants, etc.

Methods

valid_context :: Lens' t [TheoryModel]

spaces :: Lens' t [SpaceDefn]

quantities :: Lens' t [DefinedQuantityDict]

operations :: Lens' t [ConceptChunk]

defined_quant :: Lens' t [ModelQDef]

invariants :: Lens' t [ModelExpr]

defined_fun :: Lens' t [ModelQDef]

data TheoryModel

A TheoryModel is a collection of:

tUid - a UID,

con - a ConceptChunk,

vctx - definition context (TheoryModels),

spc - type definitions (SpaceDefns),

quan - quantities (DefinedQuantityDicts),

ops - operations (ConceptChunks),

defq - definitions (QDefinitions),

invs - invariants (ModelExprs),

dfun - defined functions (QDefinitions),

ref - accompanying references (DecRefs),

lb - a label (SpaceDefn),

ra - reference address (SpaceDefn),

notes - additional notes (Sentences).

Right now, neither the definition context (vctx) nor the spaces (spc) are ever defined.

tm :: (Quantity q, MayHaveUnit q, Concept q, Concept c) => ModelKind ModelExpr -> [q] -> [c] -> [ModelQDef] -> [ModelExpr] -> [ModelQDef] -> [DecRef] -> String -> [Sentence] -> TheoryModel

Constructor for theory models. Must have a source. Uses the shortname of the reference address.

tmNoRefs :: (Quantity q, MayHaveUnit q, Concept q, Concept c) => ModelKind ModelExpr -> [q] -> [c] -> [ModelQDef] -> [ModelExpr] -> [ModelQDef] -> String -> [Sentence] -> TheoryModel

Constructor for theory models. Uses the shortname of the reference address.