Theory.Drasil.InstanceModel

Description

Defines types and functions for Instance Models.

Type

data InstanceModel

An instance model is a ModelKind that may have specific inputs, outputs, and output constraints. It also has attributes like references, derivation, labels (ShortName), reference address, and notes.

Constructors

im :: ModelKind Expr -> Inputs -> Output -> OutputConstraints -> [DecRef] -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with everything defined.

imNoDeriv :: ModelKind Expr -> Inputs -> Output -> OutputConstraints -> [DecRef] -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no derivation.

imNoRefs :: ModelKind Expr -> Inputs -> Output -> OutputConstraints -> Maybe Derivation -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no references.

imNoDerivNoRefs :: ModelKind Expr -> Inputs -> Output -> OutputConstraints -> String -> [Sentence] -> InstanceModel

Smart constructor for instance models with a custom term, and no derivations or references.

Functions

getEqModQdsFromIm :: [InstanceModel] -> [SimpleQDef]

Grab all related QDefinitions from a list of instance models.

qwUC :: (Quantity q, MayHaveUnit q, Concept q) => q -> Input

For building a quantity with no constraint.

qwC :: (Quantity q, MayHaveUnit q, Concept q) => q -> RealInterval Expr Expr -> Input

For building a quantity with a constraint.