Theory.Drasil.ModelKinds

Description

Defines types and functions for creating models.

Types

data ModelKind e

ModelKinds carrier, used to carry commonly overwritten information from the IMsTMsGDs.

Constructors

MK

Fields

_mk :: ModelKinds e

_mkUID :: UID

_mkTerm :: NP

data ModelKinds e where

Models can be of different kinds:

NewDEModels represent differential equations as DifferentialModels

DEModels represent differential equations as RelationConcepts

EquationalConstraints represent invariants that will hold in a system of equations.

EquationalModels represent quantities that are calculated via a single definition/QDefinition.

EquationalRealms represent MultiDefns; quantities that may be calculated using any one of many DefiningExprs (e.g., 'x = A = ... = Z')

FunctionalModels represent quantity-resulting function definitions.

OthModels are placeholders for models. No new OthModels should be created, they should be using one of the other kinds.

Constructors

NewDEModel :: DifferentialModel -> ModelKinds e

DEModel :: RelationConcept -> ModelKinds e

EquationalConstraints :: ConstraintSet e -> ModelKinds e

EquationalModel :: QDefinition e -> ModelKinds e

EquationalRealm :: MultiDefn e -> ModelKinds e

OthModel :: RelationConcept -> ModelKinds e

Constructors

newDEModel :: String -> NP -> DifferentialModel -> ModelKind e

Smart constructor for NewDEModels

deModel :: String -> NP -> RelationConcept -> ModelKind e

Smart constructor for DEModels

equationalConstraints :: String -> NP -> ConstraintSet e -> ModelKind e

Smart constructor for EquationalConstraints

equationalModel :: String -> NP -> QDefinition e -> ModelKind e

Smart constructor for EquationalModels

equationalRealm :: String -> NP -> MultiDefn e -> ModelKind e

Smart constructor for EquationalRealms

othModel :: String -> NP -> RelationConcept -> ModelKind Expr

Smart constructor for OthModels

newDEModel' :: DifferentialModel -> ModelKind e

Smart constructor for NewDEModels, deriving UID+Term from the DifferentialModel

deModel' :: RelationConcept -> ModelKind e

Smart constructor for DEModels, deriving UID+Term from the RelationConcept

equationalConstraints' :: ConstraintSet e -> ModelKind e

Smart constructor for EquationalConstraints, deriving UID+Term from the ConstraintSet

equationalModel' :: QDefinition e -> ModelKind e

Smart constructor for EquationalModels, deriving UID+Term from the QDefinition

equationalRealm' :: MultiDefn e -> ModelKind e

Smart constructor for EquationalRealms, deriving UID+Term from the MultiDefn

othModel' :: RelationConcept -> ModelKind e

Smart constructor for OthModels, deriving UID+Term from the RelationConcept

equationalModelU :: String -> QDefinition e -> ModelKind e

Smart constructor for EquationalModels, deriving Term from the QDefinition

equationalModelN :: NP -> QDefinition e -> ModelKind e

Smart constructor for EquationalModels, deriving UID from the QDefinition

equationalRealmU :: String -> MultiDefn e -> ModelKind e

Smart constructor for EquationalRealms

equationalRealmN :: NP -> MultiDefn e -> ModelKind e

Smart constructor for EquationalRealms, deriving UID from the MultiDefn

Lenses

setMk :: ModelKinds e -> Setter' DifferentialModel a -> Setter' RelationConcept a -> Setter' (ConstraintSet e) a -> Setter' (QDefinition e) a -> Setter' (MultiDefn e) a -> a -> ModelKinds e

Map into internal representations of ModelKinds

elimMk :: Getter DifferentialModel a -> Getter RelationConcept a -> Getter (ConstraintSet e) a -> Getter (QDefinition e) a -> Getter (MultiDefn e) a -> ModelKinds e -> a

Retrieve internal data from ModelKinds

lensMk :: forall e a. Lens' DifferentialModel a -> Lens' RelationConcept a -> Lens' (ConstraintSet e) a -> Lens' (QDefinition e) a -> Lens' (MultiDefn e) a -> Lens' (ModelKinds e) a

Make a Lens for ModelKinds.

getterMk :: forall e a. Getter DifferentialModel a -> Getter RelationConcept a -> Getter (ConstraintSet e) a -> Getter (QDefinition e) a -> Getter (MultiDefn e) a -> Getter (ModelKinds e) a

Make a Getter for ModelKinds.

Functions

getEqModQds :: [ModelKind e] -> [QDefinition e]

Extract a list of QDefinitions from a list of ModelKinds.