Theory.Drasil.MultiDefn

Description

Defines types and functions for creating mult-definitions.

Types

data MultiDefn e

MultiDefns are QDefinition factories, used for showing one or more ways we can define a QDefinition.

data DefiningExpr e

DefiningExpr are the data that make up a (quantity) definition, namely the description, the defining (rhs) expression and the context domain(s). These are meant to be alternate but equivalent definitions for a single concept.

Constructors

mkMultiDefn :: String -> DefinedQuantityDict -> Sentence -> NonEmpty (DefiningExpr e) -> MultiDefn e

Smart constructor for MultiDefns, does nothing special at the moment. First argument is the Space to become a UID.

mkMultiDefnForQuant :: DefinedQuantityDict -> Sentence -> NonEmpty (DefiningExpr e) -> MultiDefn e

Smart constructor for MultiDefns defining UIDs using that of the DefinedQuantityDict.

mkDefiningExpr :: String -> [UID] -> Sentence -> e -> DefiningExpr e

Smart constructor for DefiningExprs.

Functions

multiDefnGenQD :: MultiDefn e -> DefiningExpr e -> QDefinition e

Convert MultiDefns into QDefinitions via a specific DefiningExpr.

multiDefnGenQDByUID :: MultiDefn e -> UID -> QDefinition e

Convert MultiDefns into QDefinitions via a specific DefiningExpr (by UID).